refactor: 도메인 서비스 내부 패키지 생성
