+++ backend/src/main/java/feedzupzup/backend/feedback/application/AdminFeedbackService.java
@@ -6,8 +6,8 @@ import feedzupzup.backend.feedback.domain.Feedback;
 import feedzupzup.backend.feedback.domain.FeedbackAmount;
 import feedzupzup.backend.feedback.domain.FeedbackPage;
 import feedzupzup.backend.feedback.domain.FeedbackRepository;
-import feedzupzup.backend.feedback.domain.service.FeedbackSortStrategy;
-import feedzupzup.backend.feedback.domain.service.FeedbackSortStrategyFactory;
+import feedzupzup.backend.feedback.domain.service.sort.FeedbackSortStrategy;
+import feedzupzup.backend.feedback.domain.service.sort.FeedbackSortStrategyFactory;
 import feedzupzup.backend.feedback.domain.vo.FeedbackSortBy;
 import feedzupzup.backend.feedback.domain.vo.ProcessStatus;
 import feedzupzup.backend.feedback.dto.request.UpdateFeedbackCommentRequest;
