+++ backend/src/main/java/feedzupzup/backend/feedback/application/UserFeedbackService.java
@@ -5,9 +5,9 @@ import feedzupzup.backend.category.domain.OrganizationCategory;
 import feedzupzup.backend.feedback.domain.Feedback;
 import feedzupzup.backend.feedback.domain.FeedbackPage;
 import feedzupzup.backend.feedback.domain.FeedbackRepository;
-import feedzupzup.backend.feedback.domain.service.CacheHandler;
-import feedzupzup.backend.feedback.domain.service.FeedbackSortStrategy;
-import feedzupzup.backend.feedback.domain.service.FeedbackSortStrategyFactory;
+import feedzupzup.backend.feedback.domain.service.cache.CacheHandler;
+import feedzupzup.backend.feedback.domain.service.sort.FeedbackSortStrategy;
+import feedzupzup.backend.feedback.domain.service.sort.FeedbackSortStrategyFactory;
 import feedzupzup.backend.feedback.domain.vo.FeedbackSortBy;
 import feedzupzup.backend.feedback.domain.vo.ProcessStatus;
 import feedzupzup.backend.feedback.dto.request.CreateFeedbackRequest;
