+++ backend/src/main/java/feedzupzup/backend/feedback/domain/service/cache/OldestCacheHandler.java
@@ -1,4 +1,4 @@
-package feedzupzup.backend.feedback.domain.service;
+package feedzupzup.backend.feedback.domain.service.cache;
 
 import feedzupzup.backend.feedback.dto.response.FeedbackItem;
 import java.util.UUID;
